moves keyword set to its own file - deletes test.py and moves keyword finding logic to extract.py - adds a couple todo's to scrape.py
