+++ app/keywords.py
@@ -1,0 +1,40 @@
+""" Keywords set used for keyword extraction and matching.
+This set is maintained and updated with new tech keywords.
+"""
+
+
+keywords = {
+    'aircrack-ng','aks', 'algorand', 'anaconda', 'android', 'angular', 'ansible',
+    'apache', 'appium', 'arangodb', 'arcore', 'arduino', 'arkit', 'aseprite',
+    'asp.net', 'aws', 'azure', 'azure devops', 'azure functions', 'bash', 'bigquery',
+    'bitbucket', 'blender', 'bootstrap', 'burp suite', 'c', 'c#', 'c++', 'cassandra',
+    'catboost', 'celery', 'chai', 'chainlink', 'chef', 'chrome os', 'circleci',
+    'clojure', 'cloudfront', 'cockroachdb', 'couchbase', 'cryengine', 'css',
+    'cucumber', 'cypress', 'dart', 'dask', 'django', 'docker', 'docker compose',
+    'dynamodb', 'ec2', 'elasticsearch', 'elixir', 'ember.js', 'erlang', 'ethereum',
+    'ethers.js', 'express.js', 'f#', 'fastapi', 'fastify', 'figma', 'firebase',
+    'firebase realtime database', 'firestore', 'flask', 'flutter', 'freebsd',
+    'gatsby', 'gcp', 'gin', 'git', 'github', 'github actions', 'gitlab', 'gke',
+    'go', 'godot', 'google app engine', 'google cloud platform', 'grafana',
+    'graphql', 'groovy', 'grpc', 'hadoop', 'hardhat', 'haskell', 'hasura', 'hive',
+    'html', 'hyperledger', 'iam', 'influxdb', 'ionic', 'ios', 'istio', 'jasmine',
+    'java', 'javascript', 'jenkins', 'jest', 'jira', 'jmeter', 'john the ripper',
+    'jquery', 'julia', 'jupyter', 'jwt', 'kafka', 'katalon studio', 'keras',
+    'knime', 'koa.js', 'kotlin', 'kubernetes', 'lambda', 'laravel', 'lightgbm',
+    'linux', 'loadrunner', 'lua', 'macos', 'matplotlib', 'maya', 'mercurial',
+    'metasploit', 'mocha', 'mongodb', 'mysql', 'neo4j', 'nessus', 'nestjs',
+    'next.js', 'nginx', 'nltk', 'node.js', 'nosqlmariadb', 'numpy', 'nuxt.js',
+    'oauth', 'objective-c', 'opencl', 'opencv', 'openvas', 'oracle', 'pandas',
+    'perl', 'phaser', 'phoenix', 'php', 'polkadot', 'postgresql', 'postman',
+    'power bi', 'powershell', 'prometheus', 'pulsar', 'puppet', 'puppeteer', 'pygame',
+    'python', 'pytorch', 'r', 'rabbitmq', 'rapidminer', 'rds', 'react native',
+    'redis', 'redux', 'rest', 'ruby', 'ruby on rails', 'rust', 's3', 'sass',
+    'scala', 'scikit-learn', 'scssreact', 'seaborn', 'selenium', 'snort', 'snowflake',
+    'soap', 'socket.io', 'solidity', 'spacy', 'splunk', 'spring boot', 'sql',
+    'sql server', 'sqlite', 'substance painter', 'supabase', 'svelte', 'svn',
+    'swift', 'swiftui', 'tableau', 'tailwind', 'tensorflow', 'terraform', 'testng',
+    'tezos', 'three.js', 'travis ci', 'trello', 'truffle', 'tsoa', 'typescript',
+    'unity', 'unreal engine', 'vba', 'vercel', 'visual basic', 'vite', 'vue.js',
+    'web3.js', 'webassembly', 'weka', 'windows', 'wireshark', 'xamarin', 'xgboost',
+    'xml', 'zbrush'
+    }
